Implement Authorization

+++ TodoApp/routers/todos.py
@@ -3,6 +3,7 @@ from pydantic import BaseModel, Field
 from sqlalchemy.orm import Session
 from fastapi import APIRouter, Path, Depends, HTTPException
 from starlette import status
+from .auth import get_current_user
 
 
 from models import Todos
@@ -19,6 +20,7 @@ def get_db():
         db.close()
 
 db_dependency = Annotated[Session, Depends(get_db)]
+user_dependency = Annotated[dict, Depends(get_current_user)]
 
 class TodoRequest(BaseModel):
     title: str = Field(min_length=3)
@@ -29,34 +31,44 @@ class TodoRequest(BaseModel):
 
 @router.get('/', status_code=status.HTTP_200_OK)
 # Depends means dependency injection - Need to execute something before what we are trying to execute
-async def read_all(db: db_dependency):
-    return db.query(Todos).all()
+async def read_all(user: user_dependency, db: db_dependency):
+    if not user:
+        raise HTTPException(status_code=status.HTTP_401_UNAUTHORIZED, detail='Could not validate user')
+    return db.query(Todos).filter(Todos.owner_id==user.get('id')).all()
 
 
-@router.get('/todo/{todoid}', status_code=status.HTTP_200_OK)
-async def read_todo(db: db_dependency, todo_id: int = Path(gt=0)):
-    todo_model = db.query(Todos).filter(Todos.id == todo_id).first()
+@router.get('/todo/{todo_id}', status_code=status.HTTP_200_OK)
+async def read_todo(user: user_dependency, db: db_dependency, todo_id: int = Path(gt=0)):
+    if not user:
+        raise HTTPException(status_code=status.HTTP_401_UNAUTHORIZED, detail='Could not validate user')
+    todo_model = db.query(Todos).filter(Todos.id == todo_id).filter(Todos.owner_id == user.get('id')).first()
     if todo_model is not None:
         return todo_model
-    raise HTTPException(status_code=404, detail='Todo not found')
+    raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail='Todo not found')
 
 
 @router.post('/todo', status_code=status.HTTP_201_CREATED)
-async def create_todo(db: db_dependency, todo_request: TodoRequest):
-    todo_model = Todos(**todo_request.model_dump())
+async def create_todo(user: user_dependency, db: db_dependency, todo_request: TodoRequest):
+    if user is None:
+        raise HTTPException(status_code=status.HTTP_401_UNAUTHORIZED, detail='Could not validate user')
+    todo_model = Todos(**todo_request.model_dump(), owner_id= user.get('id'))
     db.add(todo_model)
     db.commit()
 
 
 @router.put('/todo/{todo_id}', status_code=status.HTTP_200_OK)
-async def update_todo(db: db_dependency,  
+async def update_todo(user: user_dependency, 
+                      db: db_dependency,  
                       todo_request: TodoRequest,
                       todo_id: int = Path(gt=0)):
     
-    todo_model = db.query(Todos).filter(Todos.id == todo_id).first()
+    if user is None:
+        raise HTTPException(status_code=status.HTTP_401_UNAUTHORIZED, detail='Could not validate user')
+
+    todo_model = db.query(Todos).filter(Todos.id == todo_id).filter(Todos.owner_id == user.get('id')).first()
     
     if not todo_model:
-        raise HTTPException(status_code=404, detail='Todo not found')
+        raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail='Todo not found')
 
     todo_model.title = todo_request.title
     todo_model.description = todo_request.description
@@ -70,14 +82,17 @@ async def update_todo(db: db_dependency,
 
 
 @router.delete('/todo/{todo_id}', status_code=status.HTTP_204_NO_CONTENT)
-async def delete_todo(db: db_dependency, todo_id: int):
+async def delete_todo(user: user_dependency, db: db_dependency, todo_id: int):
+
+    if user is None:
+        raise HTTPException(status_code=status.HTTP_401_UNAUTHORIZED, detail='Could not validate user')
     
-    todo_model = db.query(Todos).filter(Todos.id == todo_id).first()
+    todo_model = db.query(Todos).filter(Todos.id == todo_id).filter(Todos.owner_id == user.get('id')).first()
 
     if not todo_model:
-        raise HTTPException(status_code=404, detail='Item not found')
+        raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail='Item not found')
     
-    db.query(Todos).filter(Todos.id == todo_id).delete()
+    db.query(Todos).filter(Todos.id == todo_id).filter(Todos.owner_id == user.get('id')).delete()
 
     db.commit()
